% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediccion_plaza_aparcamiento.R
\name{prediccion_plaza_aparcamiento}
\alias{prediccion_plaza_aparcamiento}
\title{Predicción del estado de un aparcamiento las proximas horas}
\usage{
prediccion_plaza_aparcamiento()
}
\value{
json
}
\description{
Modelo KNN
}
\examples{
 prediccion_plaza_aparcamiento()

}
